Move the MainWindow's logic to a View
