+++ EvoCarcassonne/MainWindow.xaml
@@ -4,31 +4,10 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:EvoCarcassonne"
+        xmlns:views = "clr-namespace:EvoCarcassonne.View"
         mc:Ignorable="d"
         Title="EvoCarcassonne" Height="450" Width="800">
     <Grid Margin="10">
-        <ItemsControl Name="Board">
-            <ItemsControl.ItemsPanel>
-                <ItemsPanelTemplate>
-                    <UniformGrid Columns="10" Rows="10" Width="{Binding ActualHeight, RelativeSource={RelativeSource Self}}" />
-                </ItemsPanelTemplate>
-            </ItemsControl.ItemsPanel>
-
-            <ItemsControl.ItemTemplate>
-                <DataTemplate>
-                    <Button Tag="{Binding Tag}" Background="{Binding Background}" Content="{Binding Content}" Click="Button_Click" BorderBrush="{x:Null}" BorderThickness="0" IsTabStop="False">
-                        <!--<Button.Style>
-                            <Style TargetType="Button">
-                                <Style.Triggers>
-                                    <Trigger Property="IsMouseOver" Value="True">
-                                        <Setter Property="Background" Value="Black"/>
-                                    </Trigger>
-                                </Style.Triggers>
-                            </Style>
-                        </Button.Style>-->
-                    </Button>
-                </DataTemplate>
-            </ItemsControl.ItemTemplate>
-        </ItemsControl>
+        <views:MainView x:Name="MainView"/>
     </Grid>
 </Window>
