+++ EvoCarcassonne/View/MainView.xaml
@@ -1,0 +1,27 @@
+﻿<UserControl x:Class="EvoCarcassonne.View.MainView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:EvoCarcassonne.View"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid>
+        <ItemsControl Name="Board">
+            <ItemsControl.ItemsPanel>
+                <ItemsPanelTemplate>
+                    <UniformGrid Columns="10" Rows="10" Width="{Binding ActualHeight, RelativeSource={RelativeSource Self}}" />
+                </ItemsPanelTemplate>
+            </ItemsControl.ItemsPanel>
+
+            <ItemsControl.ItemTemplate>
+                <DataTemplate>
+                    <Button Tag="{Binding Tag, UpdateSourceTrigger=PropertyChanged}" Background="{Binding Background, UpdateSourceTrigger=PropertyChanged}"
+                            Content="{Binding Content, UpdateSourceTrigger=PropertyChanged}"
+                            Click="Button_Click" BorderBrush="{x:Null}"
+                            BorderThickness="0" IsTabStop="False" />
+                </DataTemplate>
+            </ItemsControl.ItemTemplate>
+        </ItemsControl>
+    </Grid>
+</UserControl>
